Initial commit.
The commit contains the implementation of the thread pool.
The sample of the usage:

ThreadPool tp(std::thread::hardware_concurrency());
// wait if there are any available free thread
tp.waitAvailability();

auto task=
      make_shared<packaged_task<void()>>([](void){cout << "hello world";});
tp.push(task);

The sample runs the task in the own thread to print 'hello world'
The number of the threads are restrictes with the number of the CPU cores.

Signed-off-by: Ihor Usyk <ihor.usyk@gmail.com>